feat: [Sentry][SessionD] Add hwid tag and move to a separate file
