+++ lte/gateway/c/session_manager/SessionProxyResponderHandler.cpp
@@ -17,6 +17,7 @@
 #include "EnumToString.h"
 #include "GrpcMagmaUtils.h"
 #include "magma_logging.h"
+#include "SentryWrappers.h"
 #include "SessionProxyResponderHandler.h"
 
 using grpc::Status;
@@ -29,6 +30,7 @@ SessionProxyResponderHandlerImpl::SessionProxyResponderHandlerImpl(
 void SessionProxyResponderHandlerImpl::ChargingReAuth(
     ServerContext* context, const ChargingReAuthRequest* request,
     std::function<void(Status, ChargingReAuthAnswer)> response_callback) {
+  set_sentry_transaction("ChargingReAuth");
   auto& request_cpy = *request;
   PrintGrpcMessage(static_cast<const google::protobuf::Message&>(request_cpy));
   MLOG(MDEBUG) << "Received a Gy (Charging) ReAuthRequest for "
@@ -69,6 +71,8 @@ void SessionProxyResponderHandlerImpl::ChargingReAuth(
 void SessionProxyResponderHandlerImpl::PolicyReAuth(
     ServerContext* context, const PolicyReAuthRequest* request,
     std::function<void(Status, PolicyReAuthAnswer)> response_callback) {
+  set_sentry_transaction("PolicyReAuth");
+
   auto& request_cpy = *request;
   PrintGrpcMessage(static_cast<const google::protobuf::Message&>(request_cpy));
   MLOG(MDEBUG) << "Received a Gx (Policy) ReAuthRequest for session_id "
